whip vs grapple, beard segment animation system implemented
Whip and grapple now spawn beard segments.  Whip and grapple currently do not affect anything external to the player, but in both cases the beard segments extend out in a straight line to the target respecting max range.  Whip then causes the segments to retract again, but grapple checks for trailing segments as the player moves closer to the grapple point (not implemented yet).  Dealing damage can be handled by the beard segments themselves which can have a collider attached in the future.