health and beard length now update in the ui
Can't really test how this will work with losing health until monster is merged, but the UI now updates in response to exchanging beard and health.  Also switched the controls since it makes intuitive sense that the slider moves to the right by pressing the rightward key, which we can change later if nobody likes it.